cli: move dag to tdp command
