+++ tdp/cli/commands/service_versions.py
@@ -7,7 +7,7 @@ import click
 from tdp.core.models import ServiceLog
 from tdp.core.models.base import keyvalgen
 
-from tdp.cli.run.session import get_session_class
+from tdp.cli.session import get_session_class
 
 
 @click.command(
